+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.PartyRunApplication"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.OnBoardingActivity"
+            android:name=".presentation.auth.AuthActivity"
             android:exported="true"
             android:theme="@style/Theme.PartyRunApplication">
             <intent-filter>
@@ -27,7 +27,7 @@
         </activity>
 
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.PartyRunApplication">
         </activity>
