--- app/src/main/java/online/partyrun/partyrunapplication/network/service/AuthApiService.kt
@@ -1,4 +1,0 @@
-package online.partyrun.partyrunapplication.network.service
-
-interface AuthApiService {
-}
