--- app/src/main/java/online/partyrun/partyrunapplication/presentation/main/Test1Screen.kt
@@ -1,11 +1,0 @@
-package online.partyrun.partyrunapplication.presentation.main
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun Test1Screen(
-    navigateToTest2WithArgs: (String) -> Unit
-) {
-
-
-}
