+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/main/main1/Test1Screen.kt
@@ -1,0 +1,11 @@
+package online.partyrun.partyrunapplication.presentation.main.main1
+
+import androidx.compose.runtime.Composable
+import online.partyrun.partyrunapplication.presentation.test.TestScreen
+
+@Composable
+fun Test1Screen(
+    navigateToTest2WithArgs: (String) -> Unit
+) {
+    TestScreen()
+}
