+++ app/src/main/java/online/partyrun/partyrunapplication/presentation/main/main4/Test4Screen.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.presentation.main
+package online.partyrun.partyrunapplication.presentation.main.main4
 
 import androidx.compose.runtime.Composable
