--- app/src/main/java/online/partyrun/partyrunapplication/presentation/signin/SignInGoogleResult.kt
@@ -1,12 +1,0 @@
-package online.partyrun.partyrunapplication.presentation.signin
-
-data class SignInGoogleResult(
-    val userData: UserGoogleData?,
-    val errorMessage: String?
-)
-
-data class UserGoogleData(
-    val userId: String,
-    val username: String?,
-    val profilePictureUrl: String?
-)
