--- app/src/main/java/online/partyrun/partyrunapplication/presentation/signin/SignInGoogleState.kt
@@ -1,5 +1,0 @@
-package online.partyrun.partyrunapplication.presentation.signin
-data class SignInGoogleState(
-    val isSignInSuccessful: Boolean = false,
-    val signInError: String? = null
-)
