--- app/src/main/java/online/partyrun/partyrunapplication/presentation/signin/SignInScreen.kt
@@ -1,40 +1,0 @@
-package online.partyrun.partyrunapplication.presentation.signin
-
-
-import android.widget.Toast
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun SignInScreen(
-    state: SignInGoogleState,
-    onSignInClick: () -> Unit
-) {
-    val context = LocalContext.current
-    LaunchedEffect(key1 = state.signInError) {
-        state.signInError?.let { error ->
-            /* TODO: Toast 메세지 변경 요 */
-            Toast.makeText(context, error, Toast.LENGTH_LONG).show()
-        }
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        Button(onClick = onSignInClick) {
-            Text(text = "Sign in")
-        }
-    }
-}
