--- app/src/main/java/online/partyrun/partyrunapplication/presentation/splash/SplashScreen.kt
@@ -1,19 +1,0 @@
-package online.partyrun.partyrunapplication.presentation.splash
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import online.partyrun.partyrunapplication.presentation.components.CenterCircularProgressIndicator
-
-@Composable
-fun SplashScreen(
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-    ) {
-        CenterCircularProgressIndicator()
-    }
-}
